Build a risk calculator

Construct a formatted shiny-based risk calculator through different model inputs

Usage

risk_calculator(model, ...)

# S3 method for glm
risk_calculator(
  model,
  title = "",
  citation = "",
  label = "Predicted Value",
  label_header = "Result",
  value_header = "Value",
  format = NULL,
  app_name = NULL,
  labels = NULL,
  levels = NULL,
  placeholders = NULL,
  ...
)

# S3 method for coxph
risk_calculator(
  model,
  time,
  title = "",
  citation = "",
  label = "Survival Probability",
  label_header = "Result",
  value_header = "Value",
  format = NULL,
  app_name = NULL,
  labels = NULL,
  levels = NULL,
  placeholders = NULL,
  ...
)

Arguments

model

A glm or coxph object whose formula only has non-negated factors

...

Additional arguments

title

Title for risk calculator (see titlePanel)

citation

Citation(s) and author information

label

Label for the calculated value

label_header

Header label for the result descriptor

value_header

Header label for the result value

format

Function to format the predicted value for display

app_name

App shorthand, like "AppExample" (https://riskcalc.org/AppExample/)

labels

Named character vector specifying labels for any inputs

levels

Named list of named character vectors specifying labels for any factor levels

placeholders

Named character vector specifying range limits for numeric input variables (of the form c(var="<min>-<max>")). Invokes calls to validate enforcing the limits.

time

Single time point to calculate survival probability (see summary.survfit)

Value

A shiny application

Examples

# Make a data set
dat <- survival::gbsg
dat$meno <- factor(dat$meno)

# Build a model
mod <-
  survival::coxph(
    formula = survival::Surv(rfstime, status) ~ age + meno,
    data = dat
  )

# Time point of interest
time <- median(dat$rfstime) # 1084

# Create the risk calculator
if(interactive()) {
  risk_calculator(
    model = mod,
    time = time,
    title = "Risk of Death or Recurrence in Breast Cancer",
    citation =
      paste0(
        "Patrick Royston and Douglas Altman, External validation of a Cox prognostic",
        "model: principles and methods. BMC Medical Research Methodology 2013, 13:33"
      ),
    label_header = "Probability Of Death or Recurrence",
    label = paste0("At time = ", time),
    labels = c(age = "Age (years)", meno = "Menopausal Status"),
    levels = list(meno = c(`0` = "Premenopausal", `1` = "Postmenopausal")),
    placeholders = c(age = "21-80"),
    format = function(x) paste0(round(100 * (1 - x), 2), "%")
  )
}